pyproject.toml
[project]
name = "passos-magicos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { email = "renato.inomata@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "lightgbm>=4.6.0",
    "mlflow>=3.9.0",
    "nbformat>=5.10.4",
    "numpy>=2.4.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "xgboost>=3.2.0",
]

[project.scripts]
passos-magicos = "passos_magicos:main"

[build-system]
requires = ["uv-build>=0.8.11,<0.9.0"]
build-backend = "uv_build"

[tool.hatch.build.targets.wheel]
packages = ["src/passos_magicos"]

[dependency-groups]
dev = [
    "ipykernel>=7.2.0",
]
